Minor code reusability improvements
